retrieving data

+++ index.php
@@ -3,14 +3,12 @@ include 'db_connect.php';
 ?>
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <title>My Shop</title>
 </head>
-
 <body>
     <div class="container my-5">
         <h2 class="text-danger">List of clients</h2>
@@ -29,23 +27,35 @@ include 'db_connect.php';
                 </tr>
             </thead>
             <tbody>
-                <tr>
-                    <td>10</td>
-                    <td>Joselyne Uwingabire</td>
-                    <td>uwingajoselyne@gmail.com</td>
-                    <td>Kigali,Rwanda</td>
-                    <td>+250 785991851</td>
-                    <td>24/10/2024</td>
-                    <td>
-                        <a class="btn btn-danger btn-sm" href="/myshop/edit.php">Edit</a>
-                        <a class="btn btn-danger btn-sm" href="/myshop/delete.php">Delete</a>
-                    </td>
-                </tr>
+                <?php
+                if (!$conn) {
+                    die("Connection failed: " . $conn->error);
+                }
+                $sql = "SELECT * FROM clients";
+                $result = $conn->query($sql);
+                if (!$result) {
+                    die("Connection failed: " . $conn->error);
+                }
+                while ($row = $result->fetch_assoc()) {
+                    echo "
+                    <tr>
+                        <td>{$row['id']}</td>
+                        <td>{$row['name']}</td>
+                        <td>{$row['email']}</td>
+                        <td>{$row['address']}</td>
+                        <td>{$row['phone']}</td>
+                        <td>{$row['created_at']}</td>
+                        <td>
+                            <a class='btn btn-danger btn-sm' href='/myshop/edit.php'>Edit</a>
+                            <a class='btn btn-danger btn-sm' href='/myshop/delete.php'>Delete</a>
+                        </td>
+                    </tr>
+                    ";
+                }
+                ?>
             </tbody>
         </table>
-
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
 </body>
-
 </html>
